pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "proposal-pipeline"
version = "0.1.0"
description = "AI 교육 제안서 자동 생성 파이프라인"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "python-pptx>=1.0.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
visual = ["playwright>=1.40.0"]

[project.scripts]
proposal-generate = "proposal_pipeline.cli:main"
proposal-install-browsers = "proposal_pipeline._post_install:main"

[tool.setuptools.packages.find]
include = ["proposal_pipeline*"]

[tool.setuptools.package-data]
proposal_pipeline = [
    "templates/**/*",
    "data/**/*",
]
